+++ btrsnap/btrsnap.py
@@ -630,7 +630,7 @@ def main():
             ln -s /srv/music /snapshots/music/target
     ...         ''')
     parser.add_argument('--version', action='version',
-                        version='%(prog)s 1.1.0-dev'
+                        version='%(prog)s 1.1.0'
                         )
     subparsers = parser.add_subparsers(title='sub-commands')