+++ CHANGELOG.rst
@@ -5,8 +5,8 @@
  CHANGELOG
 -----------
 
-DEV
-~~~
+v1.1.0
+~~~~~~
 * Added sphinx code documentation
 * Cleaned up code to pep8 standards
 * Added recursive list function